Move overlay into new class
